Conflicts:
	Koulupeliprojekti/src/Level/LevelManager.cpp

+++ Koulupeliprojekti/include/Component/PhysicsComponent.h
@@ -6,18 +6,17 @@ class PhysicsComponent : public Component
 {
 public:
   PhysicsComponent();
-  PhysicsComponent(float terminal_velocity);
-  PhysicsComponent(float acceleration_x, float acceleration_y);
-  PhysicsComponent(float acceleration_x, float acceleration_y, float terminal_velocity);
+  PhysicsComponent(double terminal_velocity);
+  PhysicsComponent(double acceleration_x, double acceleration_y);
+  PhysicsComponent(double acceleration_x, double acceleration_y, double terminal_velocity);
   ~PhysicsComponent();
-  float GetAccelerationX(); // Always between 0.0 - 1.0
-  float GetAccelerationY(); // Always between 0.0 - 1.0
-  float GetTerminalVelocity();
-  // Ensures that x and y are always between 0.0 - 1.0
-  void SetAcceleration(float x, float y);
+  double GetAccelerationX();
+  double GetAccelerationY();
+  double GetTerminalVelocity();
+  void SetAcceleration(double x, double y);
 
 private:
-  float m_acceleration_x;
-  float m_acceleration_y;
-  float m_terminal_velocity;
+  double m_acceleration_x;
+  double m_acceleration_y;
+  double m_terminal_velocity;
 };
